Fix build issues after nmirror merge
1. Layering issue with original, split assembly factoring

Model was changed to depend on COR20Flags from PEReader.
Fixed by moving COR20Flags into model.

2. Types not included in merged DLL build did not implement COR20Flags

Throwing implementation added
