+++ CoreObjectModel/MutableMetadataModel/Units.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Microsoft.Cci.MetadataReader.PEFileFlags;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.Contracts;
