+++ PEReaderAndWriter/PEReader/ModuleExtensions.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+#pragma warning disable 1591 // missing doc comments
+
 using Microsoft.Cci.MetadataReader.ObjectModelImplementation;
 using System;
 using System.Collections.Generic;
