+++ PEReaderAndWriter/PEReader/PEFileStructures.cs
@@ -147,16 +147,6 @@ namespace Microsoft.Cci.MetadataReader.PEFileFlags {
     MemWrite = 0x80000000,        // Section is writeable.
   }
 
-  public enum COR20Flags : uint {
-    ILOnly = 0x00000001,
-    Bit32Required = 0x00000002,
-    ILLibrary = 0x00000004,
-    StrongNameSigned = 0x00000008,
-    NativeEntryPoint = 0x00000010,
-    TrackDebugData = 0x00010000,
-    Prefers32bits = 0x00020000,
-  }
-
   internal enum MetadataStreamKind {
     Illegal,
     Compressed,
